+++ mxcubecore/HardwareObjects/EMBL/EMBLXrayImaging.py
@@ -42,7 +42,7 @@ from mxcubecore.utils import qt_import, Colors
 from mxcubecore.TaskUtils import task
 from mxcubecore.HardwareObjects.abstract.AbstractCollect import AbstractCollect
 from mxcubecore.HardwareObjects.QtGraphicsManager import QtGraphicsManager
-from mxcubecore.HardwareObjects import queue_model_objects as qmo
+from mxcubecore.model import queue_model_objects as qmo
 from mxcubecore import HardwareRepository as HWR
