+++ mxcubecore/HardwareObjects/ESRF/ESRFMultiCollect.py
@@ -4,7 +4,7 @@ import logging
 import time
 import os
 import math
-from mxcubecore.HardwareObjects.queue_model_objects import PathTemplate
+from mxcubecore.model.queue_model_objects import PathTemplate
 from mxcubecore.utils.conversion import string_types
 from mxcubecore import HardwareRepository as HWR
