+++ mxcubecore/HardwareObjects/ESRF/ESRFSession.py
@@ -1,7 +1,7 @@
 from mxcubecore.HardwareObjects import Session
 import os
 import time
-from mxcubecore.HardwareObjects import queue_model_objects
+from mxcubecore.model import queue_model_objects
 
 
 class ESRFSession(Session.Session):
