+++ mxcubecore/HardwareObjects/ESRF/ESRFSmallXrayCentring.py
@@ -27,7 +27,7 @@ import time
 import logging
 import requests
 
-from mxcubecore.HardwareObjects import queue_model_objects
+from mxcubecore.model import queue_model_objects
 
 from mxcubecore.HardwareObjects.abstract.AbstractXrayCentring import AbstractXrayCentring
