+++ mxcubecore/HardwareObjects/abstract/AbstractXrayCentring.py
@@ -1,18 +1,18 @@
 #! /usr/bin/env python
 # encoding: utf-8
-# 
+#
 # This file is part of MXCuBE.
-# 
+#
 # MXCuBE is free software: you can redistribute it and/or modify
 # it under the terms of the GNU Lesser General Public License as published by
 # the Free Software Foundation, either version 3 of the License, or
 # (at your option) any later version.
-# 
+#
 # MXCuBE is distributed in the hope that it will be useful,
 # but WITHOUT ANY WARRANTY; without even the implied warranty of
 # MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 # GNU Lesser General Public License for more details.
-# 
+#
 # You should have received a copy of the GNU Lesser General Public License
 # along with MXCuBE.  If not, see <https://www.gnu.org/licenses/>.
 """
@@ -29,7 +29,7 @@ import abc
 
 from mxcubecore.BaseHardwareObjects import HardwareObjectYaml
 from mxcubecore import HardwareRepository as HWR
-from mxcubecore.HardwareObjects import queue_model_objects
+from mxcubecore.model import queue_model_objects
 
 
 class AbstractXrayCentring(HardwareObjectYaml):
