--- mxcubecore/HardwareObjects/datamodel/common.py
@@ -1,5 +1,0 @@
-# -*- coding: utf-8 -*-
-
-
-class ValidationError(Exception):
-    pass
