+++ test/pytest/test_procedure.py
@@ -1,5 +1,5 @@
 import gevent
-from mxcubecore.HardwareObjects import datamodel
+from mxcubecore.model import procedure_model
 from mxcubecore.HardwareObjects.abstract.AbstractProcedure import ProcedureState
 
 
@@ -13,7 +13,7 @@ def test_procedure_init(beamline):
 
 
 def test_procedure_start(beamline):
-    data = datamodel.MockDataModel(**{"exposure_time": 5})
+    data = procedure_model.MockDataModel(**{"exposure_time": 5})
     beamline.mock_procedure.start(data)
     gevent.sleep(1)
     assert beamline.mock_procedure.state == ProcedureState.BUSY
@@ -22,7 +22,7 @@ def test_procedure_start(beamline):
 
 
 def test_procedure_stop(beamline):
-    data = datamodel.MockDataModel(**{"exposure_time": 5})
+    data = procedure_model.MockDataModel(**{"exposure_time": 5})
     beamline.mock_procedure.start(data)
     gevent.sleep(1)
     assert beamline.mock_procedure.state == ProcedureState.BUSY
